Add Python test for KeyboardDrawingDialog using introspection
[CCing anaconda-devel-list to show some promising progress]

Hi Sergey,
I've finally succeeded in running gkbd_keyboard_drawing_dialog from Python
by using introspection. I think it would be nice to share it with others,
because it took me quite a lot of time to figure out how to work with the
libgnomekbd's introspection magic, so I'm sending this as a patch.
However, the introspection is very bizarre. Because gkbd_keyboard_drawing is
a prefix of gkbd_keyboard_drawing_dialog, the generated introspection has a
class named KeyboardDrawing with two constructors -- new (constructing
a new KeyboardDialog) and dialog_new (constructing a KeyboardDrawingDialog).
When invoking KeyboardDrawing() as it is common in Python with introspection,
you can't tell which one of those is invoked and thus what object you actually
get.
This is quite bizarre per se, but it is even worse. Because there is no
KeyboardDrawingDialog class in the introspection, one can only invoke dialog's
methods as e.g.
KeyboardDrawing.dialog_set_layout(dialog, registry, layout)

while with a proper introspection, it would be just
dialog.set_layout(registry, layout)

All of this is a result of having such prefixed names. I'm quite sure there
must be a way to generate introspection more properly even with such names, but
I don't know how to do it. For more details please see the Gkbd-3.0.gir file.

--
Vratislav Podzimek