libgnomekbd/gkbd-keyboard-drawing.c: draw keys with invalid keycodes
Some keys like the function key "FN" are described in geometry files but
they have no keycode associated to them. You can find such keys in the
thinkpad and in the hhk geometry files.

Allow drawing the shape of these keys with invalid keycodes, and just
skip looking for the labels as the label is derived from the keycode.

https://bugzilla.gnome.org/show_bug.cgi?id=734618