libgnomekbd/gkbd-keyboard-drawing.c: fix modifier key pressed state
When track_modifiers is true and a modifier key gets pressed, do not
free and reallocate the data and, in particular, do not reallocate
drawing->keys zeroing it; this would result in the modifier keys pressed
states to be forgotten and the modifier keys not to be shown as pressed
when in fact they are (the keyboard symbols reflect the shift level
correctly).

Fix the current situation, allowing the modifier keys pressed state to
be drawn correctly.

https://bugzilla.gnome.org/show_bug.cgi?id=734621