Add .editorconfig for easier communication of code conventions.
Many editors support .editorconfig.  From looking at some of the code
in libgnomekbd (i've really only been playing around with the C code),
it appears to be using the configuration directives I've documented
here.  The conventions documented here aren't my personally preferred
conventions, but I find it it useful to let my editor know what the
project uses, so that i can contribute without thinking too much about
whitespace, etc.

If there are other conventions, this gives people a place to add them.

Or, if the project wants to update the conventions, it provides a
place to document those changes.