Make all symbol drawings full-width, using alignment to place them
Currently, left-displayed symbols (those produced without AltGr) are
drawn on a key rendering in a full-width cell, but right-displayed
symbols (those produced with AltGr) are only drawn in half-width cells
on the key rendering.

This means that left-displayed symbols can still overlap with
right-hand symbols if they happen to be wide, but right-hand symbols
are more likely to be truncated ("ellipsized").

This change still allows the symbols to overlap, but makes
right-displayed symbols less likely to be ellipsized.  We do this by
using alignment to place right-displayed symbols to the right, instead
of using a half-width cell.