desktop: Remove obsolete Bugzilla entries
The `X-GNOME-Bugzilla-*` entries were for use by bug-buddy, a GNOME 2
technology that's been gone for over a decade. These entries are
obsolete and can be removed from the desktop file.

The desktop file then has no variables so does not need to be
configured. It is renamed from `*.in.in` to `*.in` to reflect that and
build files and POTFILES are updated for this change.